moving Modal from src/modules/modal.tsx to src/components/modal/main.tsx adding styles.ts
