+++ src/components/modal/modal.mdx
@@ -1,7 +1,7 @@
 import { Meta } from "@storybook/blocks";
-import Modal from "./modal";
+import Modal from "./main";
 import { useState } from "react";
-import { Button } from './button';
+import { Button } from '../../modules/button';
 
 export function ModalShow() {
   const [shown, setShown] = useState(false);
